Initial Commit
generatePassword() creates a randomized password of a desired length
using characters from the character array.
hashString() hashes a string using the specified algorithm.
store() is currently not implemented, but has a basic table creation for
storing passwords. This table creation (also need a table for
representing characters and their occurrence in a password) should be
moved to a separate method.