Changed Storing of Passwords and Hashes
There are now three arrays: String[] passwords, byte[][] passwordBytes,
and byte[][] hashedPasswordBytes. Unique passwords are generated
(ensured unique by comparing against those in the passwords array), each
character in each password is stored in passwordBytes, and the each byte
in the hashed password is stored in hashedPasswordBytes.
numberOfStrings, numberOfStringLengths, and startingStringLength each
describe the number of strings of each length to generate, the number of
different lengths of strings to generate, and the starting length of
string to generate from respectively.