Displays Password Stats
The character set has been reduced to the numbers 0-9.
Removed the commented out code for storing in a database.
Password stats a printed out: the iteration, the password, the length,
followed by a list (delimited by tabs) following the order: character,
occurrence in the string, and occurrence in the hashed string.